+++ src/SpinnerInfinity.tsx
@@ -1,42 +1,32 @@
 import React, { CSSProperties } from 'react';
-import { defaultProps, normalizeSize, SpinnerProps } from './helpers';
+import { secondaryColorDefaultProps, SecondaryColorSpinnerProps } from './helpers';
+import { SpinnersProps, withSharedProps } from './withSharedProps';
 
 import './SpinnerInfinity.css';
 
-export type SpinnerInfinityProps = SpinnerProps;
+export type SpinnerInfinityProps = SpinnersProps & SecondaryColorSpinnerProps;
 
-export const SpinnerInfinity = ({
-  color,
-  enabled,
-  size,
+const Component = ({
+  secondaryColor,
   speed,
   still,
   thickness,
-  style,
   ...svgProps
-}: SpinnerInfinityProps) => {
+}: SecondaryColorSpinnerProps) => {
   const strokeWidth = 7 * (thickness / 100);
-  const svgStyle: CSSProperties = {
-    color,
-    overflow: 'visible',
-    width: normalizeSize(size),
-    ...style,
-  };
   const dashStyle: CSSProperties = !still
     ? { animation: `spinners-react-infinity ${140 / speed}s linear infinite` }
     : {};
 
-  if (!enabled) return null;
-
   return (
-    <svg fill="none" viewBox="0 0 131 55" {...svgProps} style={svgStyle}>
+    <svg fill="none" viewBox="0 0 131 55" {...svgProps}>
       <defs>
         <path
           d="M46.57 45.5138C36.346 55.4954 19.8919 55.4954 9.66794 45.5138C-0.55598 35.5321 -0.55598 19.4678 9.66794 9.48624C19.8919 -0.495412 36.346 -0.495412 46.57 9.48624L84.4303 45.5138C94.6543 55.4954 111.108 55.4954 121.332 45.5138C131.556 35.5321 131.556 19.4678 121.332 9.48624C111.108 -0.495412 94.6543 -0.495412 84.4303 9.48624L46.57 45.5138Z"
           id="spinners-react-infinity-path"
         />
       </defs>
-      <use stroke="rgba(0,0,0,0.44)" strokeWidth={strokeWidth} xlinkHref="#spinners-react-infinity-path" />
+      <use stroke={secondaryColor} strokeWidth={strokeWidth} xlinkHref="#spinners-react-infinity-path" />
       <use
         fill="none"
         stroke="currentColor"
@@ -50,4 +40,7 @@ export const SpinnerInfinity = ({
     </svg>
   );
 };
-SpinnerInfinity.defaultProps = defaultProps;
+
+Component.defaultProps = secondaryColorDefaultProps;
+
+export const SpinnerInfinity = withSharedProps(Component);