+++ src/SpinnerRoundOutlined.tsx
@@ -1,57 +1,44 @@
-import React, { CSSProperties } from 'react';
-import { defaultProps, normalizeSize, SpinnerProps } from './helpers';
+import React from 'react';
+import { defaultProps, SpinnerProps } from './helpers';
+import { SpinnersProps, withSharedProps } from './withSharedProps';
 
 import './SpinnerRoundOutlined.css';
 
-export type SpinnerRoundOutlinedProps = SpinnerProps;
+const animations = [
+  {
+    r: 2,
+  },
+  {
+    name: 'spinners-react-round-outlined',
+    r: 14,
+  },
+  {
+    name: 'spinners-react-round-outlined',
+    r: 28,
+  },
+];
 
-export const SpinnerRoundOutlined = ({
-  color,
-  enabled,
-  size,
+export type SpinnerRoundOutlinedProps = SpinnersProps & SpinnerProps;
+
+const Component = ({
   speed,
   still,
   thickness,
-  style,
   ...svgProps
-}: SpinnerRoundOutlinedProps) => {
-  const svgStyle: CSSProperties = {
-    color,
-    overflow: 'visible',
-    width: normalizeSize(size),
-    ...style,
-  };
+}: SpinnerProps) => {
   const strokeWidth = 3 * (thickness / 100);
-  const animations = [
-    {
-      r: 2,
-      strokeWidth: 4,
-    },
-    {
-      name: 'spinners-react-round-outlined-inner',
-      r: 14,
-      strokeWidth,
-    },
-    {
-      name: 'spinners-react-round-outlined-outer',
-      r: 28,
-      strokeWidth,
-    },
-  ];
-
-  if (!enabled) return null;
 
   return (
-    <svg fill="none" {...svgProps} style={svgStyle} viewBox="0 0 66 66">
-      {animations.map((animation) => (
+    <svg fill="none" {...svgProps} viewBox="0 0 66 66">
+      {animations.map((animation, i) => (
         <circle
-          key={animation.name || 'still'}
+          key={`spinner-round-outlined-r${animation.r}`}
           cx="33"
           cy="33"
           fill="none"
           r={animation.r}
           stroke="currentColor"
-          strokeWidth={animation.strokeWidth}
+          strokeWidth={i ? strokeWidth : 4}
           style={animation.name && !still
             ? { animation: `${animation.name} ${140 / speed}s ease-in-out infinite` }
             : {}}
@@ -60,4 +47,7 @@ export const SpinnerRoundOutlined = ({
     </svg>
   );
 };
-SpinnerRoundOutlined.defaultProps = defaultProps;
+
+Component.defaultProps = defaultProps;
+
+export const SpinnerRoundOutlined = withSharedProps(Component);